:truck: move main page serving code
